prepare cmd package for testing
