+++ cmd/logout.go
@@ -1,7 +1,6 @@
 package cmd
 
 import (
-	"fmt"
 	"github.com/spf13/cobra"
 )
 
@@ -9,14 +8,17 @@ var logoutCmd = &cobra.Command{
 	Use:   "logout",
 	Short: "Logout of all connected clusters",
 	Long:  `Removes the tokens stored for each cluster in the configuration file.`,
-	Run: func(cmd *cobra.Command, args []string) {
-		err := ao.Logout(configLocation)
-		if err != nil {
-			fmt.Println(err)
-		}
-	},
+	RunE:  Logout,
 }
 
 func init() {
 	RootCmd.AddCommand(logoutCmd)
+}
+
+func Logout(cmd *cobra.Command, args []string) error {
+	err := ao.Logout(configLocation)
+	if err != nil {
+		return err
+	}
+	return nil
 }
