+++ cmd/pull.go
@@ -10,16 +10,20 @@ import (
 var pullCmd = &cobra.Command{
 	Use:   "pull",
 	Short: "Update local repo for AuroraConfig",
-	Run: func(cmd *cobra.Command, args []string) {
-
-		if output, err := versioncontrol.Pull(); err != nil {
-			fmt.Println(err)
-		} else {
-			fmt.Print(output)
-		}
-	},
+	RunE:  Pull,
 }
 
 func init() {
 	RootCmd.AddCommand(pullCmd)
+}
+
+func Pull(cmd *cobra.Command, args []string) error {
+
+	if output, err := versioncontrol.Pull(); err != nil {
+		return err
+	} else {
+		fmt.Print(output)
+	}
+
+	return nil
 }
