+++ cmd/save.go
@@ -3,7 +3,6 @@ package cmd
 import (
 	"fmt"
 
-	"github.com/skatteetaten/ao/pkg/command"
 	"github.com/skatteetaten/ao/pkg/versioncontrol"
 	"github.com/spf13/cobra"
 	"os"
@@ -12,16 +11,7 @@ import (
 var saveCmd = &cobra.Command{
 	Use:   "save",
 	Short: "Save changed, new and deleted files for AuroraConfig",
-	Run: func(cmd *cobra.Command, args []string) {
-		user, _ := cmd.Flags().GetString("user")
-		url := command.GetGitUrl(ao.Affiliation, user, DefaultApiClient)
-
-		if _, err := versioncontrol.Save(url, DefaultApiClient); err != nil {
-			fmt.Println(err.Error())
-		} else {
-			fmt.Println("Save success")
-		}
-	},
+	RunE:  Save,
 }
 
 func init() {
@@ -29,4 +19,17 @@ func init() {
 
 	user, _ := os.LookupEnv("USER")
 	saveCmd.Flags().StringP("user", "u", user, "Save AuroraConfig as user")
+}
+
+func Save(cmd *cobra.Command, args []string) error {
+	user, _ := cmd.Flags().GetString("user")
+	url := versioncontrol.GetGitUrl(ao.Affiliation, user, DefaultApiClient)
+
+	if _, err := versioncontrol.Save(url, DefaultApiClient); err != nil {
+		return err
+	} else {
+		fmt.Println("Save success")
+	}
+
+	return nil
 }
