+++ pkg/log/pretty_formatter.go
@@ -1,4 +1,4 @@
-package command
+package log
 
 import (
 	"bytes"
